json_mason/mason.py
from typing import List, Union, Dict
from .models import BuildingBlocks


class JSONMason:
    @staticmethod
    def build_blocks(source_data: List[Union[str, bytes]]) -> BuildingBlocks:
        """
        Parse the JSON data and build the building blocks.
        
        Parameters:
        source_data (List[Union[str, bytes]]): List of JSON strings or streams
        
        Returns:
        BuildingBlocks: The collection of building blocks
        """
        blocks = BuildingBlocks()
        # TODO: Implement parsing and block building logic
        return blocks

    @staticmethod
    def compress(source_data: List[Union[str, bytes]], blocks: BuildingBlocks) -> List[str]:
        """
        Compress the JSON data using the building blocks.
        
        Parameters:
        source_data (List[Union[str, bytes]]): List of JSON strings or streams
        blocks (BuildingBlocks): The collection of building blocks
        
        Returns:
        List[str]: List of compressed JSON strings using Jinja templates
        """
        compressed_data = []
        # TODO: Implement compression logic
        return compressed_data

    @staticmethod
    def decompress(compressed_data: List[str], blocks: BuildingBlocks) -> List[Dict]:
        """
        Decompress the JSON data using the building blocks.
        
        Parameters:
        compressed_data (List[str]): List of compressed JSON strings
        blocks (BuildingBlocks): The collection of building blocks
        
        Returns:
        List[Dict]: List of decompressed JSON objects
        """
        decompressed_data = []
        # TODO: Implement decompression logic
        return decompressed_data
